Changed the API
The Open Weather API I was using was causing lots of errors due to being
requested over HTTP. The Dark Syk API provides the data in HTTPS which
works much better.

The icons that I was using worked with codes generated by the Open
Weather API. These have now been replaced with Skycons, which are
animated and make the whole thing a little smarter.